backend/app/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.1.7
python-multipart==0.0.6
celery==5.3.4
redis==5.0.1
alembic==1.12.1
stripe==7.5.0
python-dotenv==1.0.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
email-validator==2.1.0
aiofiles==23.2.1